pyproject.toml
[tool.poetry]
name = "rtnewssearch"
version = "0.1.0"
description = ""
authors = ["Kiran <kiran.punjabi@htree.plus>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.3"
kafka-python3 = "^3.0.0"
pydantic = "^2.7.3"
bytewax = "^0.20.1"
newsdataapi = "^0.1.17"
newsapi-python = "^0.2.7"
requests = "^2.32.3"
transformers = "^4.41.2"
upstash-vector = "^0.4.0"
torch = "^2.3.1"
langchain = "^0.2.3"
pydantic-settings = "^2.3.1"
sentence-transformers = "^3.0.0"
streamlit = "^1.35.0"

[tool.poetry.group.dev.dependencies]
confluent-kafka = "^2.4.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
